Fix some warnings and progress towards refactor
Fixed 2 bugs:
1) Don't let us discard our own princess using a prince (which would
cause an auto-lose, instead of nothing as it previously computed)
2) Fix calculation for determining how many actual decks we need to
evaluate. While the number of permutations is n! (where n is the number
of cards in the deck), we don't need to consider each permutation
seperately. For example, if we have 2 guards remaining in the deck,
called G1 and G2, it doesn't matter if G1 or G2 is the burn card, so
long as a "guard" is the burn card (for the purposes of the minmax
search). In that case, we can divide the number of permutations that are
possible by counting all repeats of cards in a remaining deck and
factorializing that. IE: if we have 2 guards and 2 handmaids in a deck
of 10, each given permutation can be made 2! * 2! (or 4) ways, which
means our total size is 10! / 4 = 907,200 decks to actually consider.
This doesn't change our math, as it turns out, only our progress
counter.

Some minor refactorings to start movnig game logic out of minimax and
into a seperate file as part of making things runnable with different
search algorithms

+++ CardDeck.cpp
@@ -51,7 +51,7 @@ Cards CardDeck::getPossibleCards(Card myCard, const Cards& myPlayedCards, const 
 
 float CardDeck::getWinProbality(Card myCard, const Cards & theirPossibleCards)
 {
-	const unsigned int totalPossibilities = theirPossibleCards.size();
+	const size_t totalPossibilities = theirPossibleCards.size();
 	unsigned int totalWins = 0;
 
 	for (const Card& possibleCard : theirPossibleCards)
